Move temp idToken console.log to My Profile
