+++ frontend/admin/src/components/Menu.tsx
@@ -19,7 +19,7 @@ import { User, UserRole } from 'shared/models'
 const useMenuItems = (user?: User) => {
   const ProfileMenuItemLink = () => (
     <MenuItemLink
-      to={user ? `/user/${user.id}` : ''}
+      to={user ? `/user/${user.id}/show` : ''}
       primaryText="My Profile"
       leftIcon={<ProfileIcon />}
     />
