+++ frontend/admin/src/components/User/UserShow.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 import {
   EmailField,
   ReferenceField,
@@ -11,12 +11,27 @@ import {
 import { Typography } from '@mui/material'
 import { User } from 'shared/models'
 
+import { getIdToken } from '../../authProvider'
+
 const UserTitle = ({ record }: { record?: User }) => (
   <span>{record ? record.name : 'User'}</span>
 )
 
 const UserShow: (props: ShowProps) => JSX.Element = (props) => {
   const { identity } = useGetIdentity()
+
+  // TEMP console.log firebase idToken for dev purposes
+  useEffect(() => {
+    const logIdToken = async () => {
+      // eslint-disable-next-line no-console
+      console.log(await getIdToken())
+    }
+
+    if (process.env.NODE_ENV === 'development') {
+      logIdToken()
+    }
+  }, [])
+
   return identity ? (
     <Show {...props} title={<UserTitle />}>
       {identity.tenantId === undefined ? (
